Some minor fixes and added a "cautious" flag.
This flag turns off any change that can break valid code.  So far
it just turns off striping the inside of parenthesis ie normally
( a == b ) becomes (a == b) without cautious.  This can cause a problem
if you have a multiline boolean expression and have #ifdef/#endif's
within the parenthesis.  I found an example where stripping would bring
the closing ) up to the same line as an #endif which breaks it (I
guess all characters on the same line as a # statement are hidden)